requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.7.1=pypi_0
aiohttp=3.5.4=pypi_0
alabaster=0.7.12=py36_0
allennlp=0.8.5=pypi_0
apex=0.1=pypi_0
argh=0.26.2=pypi_0
asn1crypto=1.2.0=py36_0
astor=0.8.0=pypi_0
astroid=1.6.6=pypi_0
async-timeout=3.0.1=pypi_0
atari-py=0.2.6=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.1.0=pypi_0
babel=2.7.0=py_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=pypi_0
blis=0.2.4=pypi_0
boto3=1.9.216=pypi_0
botocore=1.12.216=pypi_0
ca-certificates=2019.10.16=0
certifi=2019.9.11=py36_0
cffi=1.13.1=py36h2e261b9_0
chardet=3.0.4=py36_1003
click=7.0=pypi_0
cloudpickle=1.2.1=pypi_0
codecov=2.0.15=pypi_0
conda-pack=0.4.0=py_0
conllu=1.3.1=pypi_0
coverage=4.5.4=pypi_0
cryptography=2.8=py36h1ba5d50_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
cymem=2.0.2=pypi_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
docutils=0.15.2=py36_0
editdistance=0.5.3=pypi_0
en-core-web-sm=2.1.0=pypi_0
expat=2.2.6=he6710b0_0
fastcache=1.1.0=py36h7b6447c_0
flaky=3.6.1=pypi_0
flask=1.1.1=pypi_0
flask-cors=3.0.8=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
ftfy=5.6=pypi_0
future=0.17.1=pypi_0
gast=0.2.2=pypi_0
gevent=1.4.0=pypi_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py36h10f8cd9_2
greenlet=0.4.15=pypi_0
grpcio=1.23.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gym=0.14.0=pypi_0
h5py=2.9.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
idna-ssl=1.1.0=pypi_0
imagesize=1.1.0=pypi_0
importlib-metadata=0.19=pypi_0
intel-openmp=2019.4=243
ipython=7.7.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
isort=4.3.21=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.15.1=py36_0
jinja2=2.10.1=pypi_0
jmespath=0.9.4=pypi_0
joblib=0.13.2=pypi_0
jpeg=9b=h024ee3a_2
jsonnet=0.13.0=pypi_0
jsonpickle=1.2=pypi_0
kiwisolver=1.1.0=py36he6710b0_0
lazy-object-proxy=1.4.2=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
livereload=2.6.1=pypi_0
llvmlite=0.29.0=py36hd408876_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.1=pypi_0
mccabe=0.6.1=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
more-itertools=7.2.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpi4py=3.0.2=pypi_0
mpmath=1.1.0=py36_0
multidict=4.5.2=pypi_0
murmurhash=1.0.2=pypi_0
mypy=0.521=pypi_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
nltk=3.4.5=pypi_0
numba=0.45.1=py36h962f231_0
numpy=1.16.2=pypi_0
numpydoc=0.9.1=py_0
olefile=0.46=py36_0
opencv-python=4.1.1.26=pypi_0
openssl=1.1.1d=h7b6447c_3
overrides=1.9=pypi_0
packaging=19.1=pypi_0
pandas=0.25.1=pypi_0
parsimonious=0.8.1=pypi_0
parso=0.5.1=py_0
pathtools=0.1.2=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pillow=6.2.0=py36h34e0f95_0
pip=19.3.1=py36_0
pkginfo=1.5.0.1=pypi_0
plac=0.9.6=pypi_0
pluggy=0.12.0=pypi_0
port-for=0.3.1=pypi_0
preshed=2.0.1=pypi_0
prompt_toolkit=2.0.10=py_0
protobuf=3.9.1=pypi_0
ptyprocess=0.6.0=py36_0
py=1.8.0=pypi_0
pybullet=2.5.5=pypi_0
pycparser=2.19=py36_0
pyglet=1.3.2=pypi_0
pygments=2.4.2=py_0
pylint=1.9.4=pypi_0
pyopenssl=19.0.0=py36_0
pypandoc=1.4=pypi_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36_0
pytest=5.1.1=pypi_0
pytest-cov=2.7.1=pypi_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
pytorch=1.3.0=py3.6_cuda10.0.130_cudnn7.6.3_0
pytorch-pretrained-bert=0.6.2=pypi_0
pytorch-transformers=1.1.0=pypi_0
pytz=2019.3=py_0
pyyaml=5.1.2=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
readme-renderer=24.0=pypi_0
regex=2019.8.19=pypi_0
requests=2.22.0=py36_0
requests-toolbelt=0.9.1=pypi_0
responses=0.10.6=pypi_0
s3transfer=0.2.1=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.1=pypi_0
sentencepiece=0.1.83=pypi_0
setuptools=39.1.0=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=pypi_0
snowballstemmer=1.9.0=pypi_0
spacy=2.1.8=pypi_0
sphinx=2.2.1=py_0
sphinx-autobuild=0.7.1=pypi_0
sphinx-rtd-theme=0.4.3=pypi_0
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
sqlite=3.30.1=h7b6447c_0
sqlparse=0.3.0=pypi_0
srsly=0.1.0=pypi_0
stable-baselines=2.7.0=pypi_0
sympy=1.4=py36_0
tbb=2019.8=hfd86e86_0
tensorboard=1.10.0=pypi_0
tensorboardx=1.8=pypi_0
tensorflow=1.10.0=pypi_0
termcolor=1.1.0=pypi_0
thinc=7.0.8=pypi_0
timeout-decorator=0.4.1=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.4.1=py36_cu100
tornado=6.0.3=py36h7b6447c_0
tqdm=4.35.0=pypi_0
traitlets=4.3.3=py36_0
twine=1.13.0=pypi_0
typed-ast=1.0.4=pypi_0
typing-extensions=3.7.4=pypi_0
ujson=1.35=py36h14c3975_0
unidecode=1.1.1=pypi_0
urllib3=1.25.3=pypi_0
wasabi=0.2.2=pypi_0
watchdog=0.9.0=pypi_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.15.5=pypi_0
wheel=0.33.6=py36_0
word2number=1.1=pypi_0
wrapt=1.11.2=pypi_0
xz=5.2.4=h14c3975_4
yarl=1.3.0=pypi_0
zipp=0.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
